Add personalized real kind
Thank to the new module m_kind.F90 we are able in the compilation to select the real kind wanted thanks to the flag : -DKIND_REAL=something

New kind can be designed in this module. For now only two of them exist : sp and dp

They are defined with the function : select_real_kind(a,b) that normally allow o have a portable real kind on any devices.

A tidious work of rename real into real(mp) has been done in all sources to allowed this change.